+++ CHANGELOG.md
@@ -3,6 +3,8 @@ main
 
 Improvements:
 - [Issue #143]: `ctrl-z` will now send jless to the background
+- `:w[rite] <file>` and `:w[rite]! <file>` can be used to write the
+  current input to a file
 
 v0.9.0 (2023-07-16)
 ==================